Move `PodIdx` to `pos-idx.hh` and `PosTable` to `pos-table.hh`
